refactor(guest-agent): split checkpoint integration tests
