+++ crates/guest-agent/tests/integration/checkpoint/mod.rs
@@ -1,0 +1,3 @@
+mod recovery;
+mod success;
+mod support;
