+++ crates/guest-agent/tests/integration/checkpoint/success.rs
@@ -1,3 +1,4 @@
+use super::support::*;
 use crate::support::*;
 use guest_contracts::session_history_identity::{
     FinalSessionHistoryFramework, FinalSessionHistoryIdentity, FinalSessionHistoryRefKind,
@@ -7,348 +8,13 @@ use serde_json::json;
 use sha2::{Digest, Sha256};
 #[cfg(target_os = "linux")]
 use std::{
-    ffi::CString,
     fs::OpenOptions,
-    os::unix::{ffi::OsStrExt, fs::OpenOptionsExt},
+    io::Write,
+    os::unix::fs::OpenOptionsExt,
     sync::mpsc,
     thread,
     time::{Duration, Instant},
 };
-use std::{
-    ffi::OsString,
-    io::{Seek, SeekFrom, Write},
-};
-const LARGE_SESSION_HISTORY_SIZE_BYTES: usize = 1024 * 1024 + 1;
-
-fn runtime_from_process_env() -> Result<guest_agent::run_context::GuestRuntime, String> {
-    guest_agent::run_context::GuestRuntime::from_process_env()
-}
-
-fn set_claude_session_pruning(runtime: &mut guest_agent::run_context::GuestRuntime, enabled: bool) {
-    runtime
-        .config
-        .feature_flags
-        .insert("claudeSessionPruning".to_string(), enabled);
-}
-
-fn session_file_paths() -> (String, String) {
-    let paths = shared_guest_paths();
-    (
-        paths.session_id_file().to_string(),
-        paths.session_history_path_file().to_string(),
-    )
-}
-
-struct EnvVarRestore {
-    key: &'static str,
-    value: Option<OsString>,
-}
-
-impl EnvVarRestore {
-    fn capture(key: &'static str) -> Self {
-        Self {
-            key,
-            value: std::env::var_os(key),
-        }
-    }
-}
-
-impl Drop for EnvVarRestore {
-    fn drop(&mut self) {
-        unsafe {
-            match &self.value {
-                Some(value) => std::env::set_var(self.key, value),
-                None => std::env::remove_var(self.key),
-            }
-        }
-    }
-}
-
-fn write_derived_claude_history(
-    home_dir: &str,
-    session_id: &str,
-    history: &str,
-) -> Result<(), String> {
-    let (session_id_file, _) = session_file_paths();
-    guest_agent::paths::write_private(&session_id_file, session_id)
-        .map_err(|e| format!("write session id: {e}"))?;
-    let project_name = guest_agent::paths::CANONICAL_WORKING_DIR
-        .strip_prefix('/')
-        .unwrap_or(guest_agent::paths::CANONICAL_WORKING_DIR)
-        .replace('/', "-");
-    let history_path = std::path::Path::new(home_dir)
-        .join(".claude")
-        .join("projects")
-        .join(format!("-{project_name}"))
-        .join(format!("{session_id}.jsonl"));
-    let parent = history_path
-        .parent()
-        .ok_or_else(|| format!("history path has no parent: {}", history_path.display()))?;
-    std::fs::create_dir_all(parent)
-        .map_err(|e| format!("create history dir {}: {e}", parent.display()))?;
-    std::fs::write(&history_path, history)
-        .map_err(|e| format!("write history {}: {e}", history_path.display()))?;
-    Ok(())
-}
-
-fn write_literal_session_history(
-    session_id: &str,
-    history: &[u8],
-) -> Result<tempfile::TempDir, String> {
-    let (session_id_file, session_history_path_file) = session_file_paths();
-    let dir = tempfile::tempdir().map_err(|e| format!("create temp history dir: {e}"))?;
-    let history_path = dir.path().join(format!("{session_id}.jsonl"));
-    std::fs::write(&history_path, history)
-        .map_err(|e| format!("write history {}: {e}", history_path.display()))?;
-    guest_agent::paths::write_private(&session_id_file, session_id)
-        .map_err(|e| format!("write session id: {e}"))?;
-    guest_agent::paths::write_private(
-        &session_history_path_file,
-        history_path.to_string_lossy().as_ref(),
-    )
-    .map_err(|e| format!("write session history marker: {e}"))?;
-    Ok(dir)
-}
-
-#[cfg(target_os = "linux")]
-fn create_fifo(path: &std::path::Path) -> std::io::Result<()> {
-    let path = CString::new(path.as_os_str().as_bytes())
-        .map_err(|error| std::io::Error::new(std::io::ErrorKind::InvalidInput, error))?;
-    let result = unsafe { libc::mkfifo(path.as_ptr(), 0o600) };
-    if result == 0 {
-        Ok(())
-    } else {
-        Err(std::io::Error::last_os_error())
-    }
-}
-
-fn write_prunable_claude_history(session_id: &str) -> Result<(tempfile::TempDir, Vec<u8>), String> {
-    let boundary_id = "bbbbbbbb-bbbb-4bbb-8bbb-bbbbbbbbbbbb";
-    let summary_id = "cccccccc-cccc-4ccc-8ccc-cccccccccccc";
-    let records = [
-        json!({
-            "type": "system",
-            "subtype": "compact_boundary",
-            "sessionId": session_id,
-            "uuid": boundary_id,
-            "parentUuid": null,
-            "logicalParentUuid": "11111111-1111-4111-8111-111111111111",
-            "isSidechain": false,
-            "version": "2.1.220"
-        }),
-        json!({
-            "type": "user",
-            "sessionId": session_id,
-            "uuid": summary_id,
-            "parentUuid": boundary_id,
-            "isCompactSummary": true,
-            "message": {"role": "user", "content": "retained summary"}
-        }),
-    ];
-    let mut candidate = Vec::new();
-    for record in records {
-        let mut line =
-            serde_json::to_vec(&record).map_err(|error| format!("encode history: {error}"))?;
-        line.push(b'\n');
-        candidate.extend_from_slice(&line);
-    }
-
-    let history_dir =
-        tempfile::tempdir().map_err(|error| format!("create history dir: {error}"))?;
-    let history_path = history_dir.path().join(format!("{session_id}.jsonl"));
-    let mut history_file =
-        std::fs::File::create(&history_path).map_err(|error| format!("create history: {error}"))?;
-    history_file
-        .set_len(guest_session_prune::CLAUDE_COMPACT_GENERATION_MAX_BYTES + 1)
-        .map_err(|error| format!("size history: {error}"))?;
-    history_file
-        .seek(SeekFrom::End(0))
-        .map_err(|error| format!("seek history: {error}"))?;
-    history_file
-        .write_all(b"\n")
-        .and_then(|()| history_file.write_all(&candidate))
-        .and_then(|()| history_file.flush())
-        .map_err(|error| format!("write history: {error}"))?;
-
-    let (session_id_file, session_history_path_file) = session_file_paths();
-    guest_agent::paths::write_private(&session_id_file, session_id)
-        .map_err(|error| format!("write session id: {error}"))?;
-    guest_agent::paths::write_private(
-        &session_history_path_file,
-        history_path.to_string_lossy().as_ref(),
-    )
-    .map_err(|error| format!("write history marker: {error}"))?;
-    Ok((history_dir, candidate))
-}
-
-fn write_prunable_codex_history(
-    session_id: &str,
-) -> Result<(tempfile::TempDir, std::path::PathBuf, Vec<u8>), String> {
-    let history_dir =
-        tempfile::tempdir().map_err(|error| format!("create Codex history dir: {error}"))?;
-    let day_dir = history_dir
-        .path()
-        .join(".codex")
-        .join("sessions")
-        .join("2026")
-        .join("07")
-        .join("26");
-    std::fs::create_dir_all(&day_dir)
-        .map_err(|error| format!("create Codex session directory: {error}"))?;
-    let history_path = day_dir.join(format!("rollout-{session_id}.jsonl"));
-
-    let line = |record_type: &str, payload: serde_json::Value| -> Result<Vec<u8>, String> {
-        let mut bytes = serde_json::to_vec(&json!({
-            "timestamp": "2026-07-26T00:00:00Z",
-            "type": record_type,
-            "payload": payload,
-        }))
-        .map_err(|error| format!("encode Codex history: {error}"))?;
-        bytes.push(b'\n');
-        Ok(bytes)
-    };
-    let canonical = line(
-        "session_meta",
-        json!({
-            "id": session_id,
-            "session_id": session_id,
-            "timestamp": "2026-07-26T00:00:00Z",
-            "cwd": guest_agent::paths::CANONICAL_WORKING_DIR,
-            "originator": "codex",
-            "cli_version": "0.144.6",
-            "source": "cli",
-            "history_mode": "legacy",
-        }),
-    )?;
-    let turn_id = "compact-turn";
-    let retained_records = [
-        line(
-            "event_msg",
-            json!({
-                "type": "task_started",
-                "turn_id": turn_id,
-                "model_context_window": 258400,
-                "collaboration_mode_kind": "default",
-            }),
-        )?,
-        line(
-            "event_msg",
-            json!({
-                "type": "user_message",
-                "message": "compact this thread",
-                "images": [],
-                "local_images": [],
-                "audio": [],
-                "local_audio": [],
-                "text_elements": [],
-            }),
-        )?,
-        line(
-            "turn_context",
-            json!({
-                "turn_id": turn_id,
-                "cwd": guest_agent::paths::CANONICAL_WORKING_DIR,
-                "approval_policy": {"granular": {"sandbox_approval": true}},
-                "sandbox_policy": {"type": "read_only"},
-                "model": "gpt-test",
-                "summary": "auto",
-            }),
-        )?,
-        line(
-            "compacted",
-            json!({
-                "message": "retained summary",
-                "replacement_history": [{
-                    "type": "message",
-                    "role": "user",
-                    "content": [{
-                        "type": "input_text",
-                        "text": "retained summary",
-                    }],
-                }],
-                "window_number": 1,
-                "first_window_id": "019c0000-0000-7000-8000-000000000001",
-                "window_id": "019c0000-0000-7000-8000-000000000002",
-            }),
-        )?,
-        line(
-            "world_state",
-            json!({"full": true, "state": {"working_directory": "/workspace"}}),
-        )?,
-        line(
-            "event_msg",
-            json!({
-                "type": "task_complete",
-                "turn_id": turn_id,
-                "last_agent_message": "retained summary",
-            }),
-        )?,
-    ];
-    let candidate = std::iter::once(canonical.clone())
-        .chain(retained_records.iter().cloned())
-        .flatten()
-        .collect::<Vec<_>>();
-
-    let mut history_file = std::fs::File::create(&history_path)
-        .map_err(|error| format!("create Codex history: {error}"))?;
-    history_file
-        .write_all(&canonical)
-        .map_err(|error| format!("write canonical Codex history: {error}"))?;
-    history_file
-        .set_len(api_contracts::generated::constants::runners::RESUME_SESSION_HISTORY_MAX_BYTES + 1)
-        .map_err(|error| format!("size Codex history: {error}"))?;
-    history_file
-        .seek(SeekFrom::End(0))
-        .and_then(|_| history_file.write_all(b"\n"))
-        .and_then(|()| {
-            for record in retained_records {
-                history_file.write_all(&record)?;
-            }
-            history_file.flush()
-        })
-        .map_err(|error| format!("write retained Codex history: {error}"))?;
-
-    let (session_id_file, _) = session_file_paths();
-    guest_agent::paths::write_private(&session_id_file, session_id)
-        .map_err(|error| format!("write Codex session id: {error}"))?;
-    Ok((history_dir, history_path, candidate))
-}
-
-fn zstd_session_history_for_test(history: &[u8]) -> std::io::Result<Vec<u8>> {
-    zstd::stream::encode_all(history, 3)
-}
-
-fn high_entropy_history(size: usize) -> Vec<u8> {
-    let mut state = 0x6a09e667f3bcc909_u64;
-    let mut bytes = Vec::with_capacity(size);
-    for _ in 0..size {
-        state ^= state >> 12;
-        state ^= state << 25;
-        state ^= state >> 27;
-        let value = state.wrapping_mul(0x2545f4914f6cdd1d);
-        bytes.push((value >> 56) as u8);
-    }
-    bytes
-}
-
-fn upload_zstd_validation_response(
-    req: &HttpMockRequest,
-    expected_body: &[u8],
-) -> HttpMockResponse {
-    if request_header_absent(req, "authorization")
-        && request_header_absent(req, "x-vercel-protection-bypass")
-        && req.body_ref() != expected_body
-        && zstd::stream::decode_all(req.body_ref()).is_ok_and(|decoded| decoded == expected_body)
-    {
-        http_status(200)
-    } else {
-        http_status(400)
-    }
-}
-
-// =========================================================================
-// Success checkpoint
-// =========================================================================
 
 #[cfg(target_os = "linux")]
 #[tokio::test(flavor = "current_thread")]
@@ -1472,413 +1138,4 @@ async fn success_checkpoint_uses_explicit_runtime_after_process_env_changes() {
         !std::path::Path::new(stale_paths.final_session_history_identity_file()).exists(),
         "stale process env runtime path must not receive final identity"
     );
-}
-
-// =========================================================================
-// Recovery checkpoint
-// =========================================================================
-
-#[tokio::test]
-async fn recovery_checkpoint_uploads_valid_session_history() {
-    let api = SharedApiMock::new().await;
-    let server = api.server();
-
-    let runtime = runtime_from_process_env().unwrap();
-    let _files_guard = SessionCheckpointFilesGuard::new();
-    let dir = tempfile::tempdir().unwrap();
-    let history_path = dir.path().join("history.jsonl");
-    let history = r#"{"type":"system"}"#.to_string() + "\n" + r#"{"type":"assistant"}"# + "\n";
-    std::fs::write(&history_path, &history).unwrap();
-    let (session_id_file, session_history_path_file) = session_file_paths();
-    guest_agent::paths::write_private(&session_id_file, "recovery-session").unwrap();
-    guest_agent::paths::write_private(
-        &session_history_path_file,
-        history_path.to_string_lossy().as_ref(),
-    )
-    .unwrap();
-
-    let prepare_mock = server.mock(|when, then| {
-        when.method(POST)
-            .path("/api/webhooks/agent/checkpoints/prepare-history")
-            .json_body_includes(r#"{"runId":"test-run-001"}"#);
-        then.status(200)
-            .header("Content-Type", "application/json")
-            .json_body(json!({
-                "presignedUrl": server.url("/test/recovery-history-upload"),
-                "existing": false
-            }));
-    });
-    let upload_mock = server.mock(|when, then| {
-        when.method(PUT)
-            .path("/test/recovery-history-upload")
-            .header("Content-Type", "application/octet-stream")
-            .body(history.as_str());
-        then.status(200);
-    });
-    let checkpoint_mock = server.mock(|when, then| {
-        when.method(POST)
-            .path("/api/webhooks/agent/checkpoints")
-            .json_body_includes(r#"{"cliAgentSessionId":"recovery-session"}"#);
-        then.status(200)
-            .header("Content-Type", "application/json")
-            .json_body(json!({"checkpointId": "checkpoint-recovery"}));
-    });
-
-    let result = guest_agent::checkpoint::create_recovery_checkpoint_for_runtime(&runtime).await;
-
-    assert!(result.is_ok());
-    prepare_mock.assert_calls_async(1).await;
-    upload_mock.assert_calls_async(1).await;
-    checkpoint_mock.assert_calls_async(1).await;
-    assert!(
-        !std::path::Path::new(runtime.paths.final_session_history_identity_file()).exists(),
-        "recovery checkpoint must not write final session history identity metadata"
-    );
-}
-
-#[tokio::test]
-async fn recovery_checkpoint_does_not_prune_eligible_claude_history() {
-    let api = SharedApiMock::new().await;
-    let server = api.server();
-
-    let mut runtime = runtime_from_process_env().unwrap();
-    set_claude_session_pruning(&mut runtime, true);
-    let _files_guard = SessionCheckpointFilesGuard::new();
-    let session_id = "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa";
-    let (history_dir, _) = write_prunable_claude_history(session_id).unwrap();
-    let history_path = history_dir.path().join(format!("{session_id}.jsonl"));
-    let source_size = std::fs::metadata(&history_path).unwrap().len();
-
-    let prepare_mock = server.mock(|when, then| {
-        when.method(POST)
-            .path("/api/webhooks/agent/checkpoints/prepare-history");
-        then.status(200);
-    });
-    let checkpoint_mock = server.mock(|when, then| {
-        when.method(POST).path("/api/webhooks/agent/checkpoints");
-        then.status(200);
-    });
-
-    let error = guest_agent::checkpoint::create_recovery_checkpoint_for_runtime(&runtime)
-        .await
-        .unwrap_err();
-
-    assert!(
-        error
-            .to_string()
-            .contains("Session history line 1 is not valid JSON"),
-        "recovery checkpoint should validate the original history: {error}"
-    );
-    assert_eq!(std::fs::metadata(&history_path).unwrap().len(), source_size);
-    assert!(!std::path::Path::new(runtime.paths.final_session_history_identity_file()).exists());
-    prepare_mock.assert_calls_async(0).await;
-    checkpoint_mock.assert_calls_async(0).await;
-}
-
-#[tokio::test]
-async fn recovery_checkpoint_does_not_prune_eligible_codex_history() {
-    let api = SharedApiMock::new().await;
-    let server = api.server();
-
-    let mut runtime = runtime_from_process_env().unwrap();
-    runtime.config.framework = guest_agent::env::Framework::Codex;
-    let _files_guard = SessionCheckpointFilesGuard::new();
-    let session_id = "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa";
-    let (history_dir, history_path, _) = write_prunable_codex_history(session_id).unwrap();
-    runtime.config.home_dir = history_dir.path().to_string_lossy().into_owned();
-    let source_size = std::fs::metadata(&history_path).unwrap().len();
-
-    let prepare_mock = server.mock(|when, then| {
-        when.method(POST)
-            .path("/api/webhooks/agent/checkpoints/prepare-history");
-        then.status(200);
-    });
-    let checkpoint_mock = server.mock(|when, then| {
-        when.method(POST).path("/api/webhooks/agent/checkpoints");
-        then.status(200);
-    });
-
-    let error = guest_agent::checkpoint::create_recovery_checkpoint_for_runtime(&runtime)
-        .await
-        .unwrap_err();
-
-    assert!(
-        error
-            .to_string()
-            .contains("Session history exceeds maximum size"),
-        "recovery checkpoint must retain the original Codex hard-limit behavior: {error}"
-    );
-    assert_eq!(std::fs::metadata(&history_path).unwrap().len(), source_size);
-    assert!(!std::path::Path::new(runtime.paths.final_session_history_identity_file()).exists());
-    prepare_mock.assert_calls_async(0).await;
-    checkpoint_mock.assert_calls_async(0).await;
-}
-
-async fn assert_recovery_checkpoint_derives_claude_history_marker(
-    seed_empty_marker: bool,
-    upload_path: &str,
-) -> Result<(), String> {
-    let api = SharedApiMock::new().await;
-    let server = api.server();
-
-    let runtime = runtime_from_process_env()?;
-    let _files_guard = SessionCheckpointFilesGuard::new();
-    let session_id = if seed_empty_marker {
-        "derived-empty-marker-session"
-    } else {
-        "derived-missing-marker-session"
-    };
-    let history = r#"{"type":"system"}"#.to_string() + "\n" + r#"{"type":"assistant"}"# + "\n";
-    if seed_empty_marker {
-        let (_, session_history_path_file) = session_file_paths();
-        guest_agent::paths::write_private(&session_history_path_file, "")
-            .map_err(|e| format!("write empty history marker: {e}"))?;
-    }
-    write_derived_claude_history(&runtime.config.home_dir, session_id, &history)?;
-
-    let prepare_mock = server.mock(|when, then| {
-        when.method(POST)
-            .path("/api/webhooks/agent/checkpoints/prepare-history")
-            .json_body_includes(r#"{"runId":"test-run-001"}"#);
-        then.status(200)
-            .header("Content-Type", "application/json")
-            .json_body(json!({
-                "presignedUrl": server.url(upload_path),
-                "existing": false
-            }));
-    });
-    let upload_mock = server.mock(|when, then| {
-        when.method(PUT)
-            .path(upload_path)
-            .header("Content-Type", "application/octet-stream")
-            .body(history.as_str());
-        then.status(200);
-    });
-    let checkpoint_mock = server.mock(|when, then| {
-        when.method(POST)
-            .path("/api/webhooks/agent/checkpoints")
-            .json_body_includes(format!(r#"{{"cliAgentSessionId":"{session_id}"}}"#));
-        then.status(200)
-            .header("Content-Type", "application/json")
-            .json_body(json!({"checkpointId": "checkpoint-derived-history"}));
-    });
-
-    let result = guest_agent::checkpoint::create_recovery_checkpoint_for_runtime(&runtime).await;
-
-    assert!(result.is_ok());
-    prepare_mock.assert_calls_async(1).await;
-    upload_mock.assert_calls_async(1).await;
-    checkpoint_mock.assert_calls_async(1).await;
-    Ok(())
-}
-
-#[tokio::test]
-async fn recovery_checkpoint_derives_missing_claude_history_marker() {
-    assert_recovery_checkpoint_derives_claude_history_marker(
-        false,
-        "/test/derived-missing-history-upload",
-    )
-    .await
-    .unwrap();
-}
-
-#[tokio::test]
-async fn recovery_checkpoint_derives_empty_claude_history_marker() {
-    assert_recovery_checkpoint_derives_claude_history_marker(
-        true,
-        "/test/derived-empty-history-upload",
-    )
-    .await
-    .unwrap();
-}
-
-#[tokio::test]
-async fn recovery_checkpoint_rejects_partial_jsonl_without_error_file() {
-    let api = SharedApiMock::new().await;
-    let server = api.server();
-
-    let runtime = runtime_from_process_env().unwrap();
-    let _files_guard = SessionCheckpointFilesGuard::new();
-    let dir = tempfile::tempdir().unwrap();
-    let history_path = dir.path().join("partial.jsonl");
-    std::fs::write(
-        &history_path,
-        r#"{"type":"system"}"#.to_string() + "\n" + r#"{"type":"assistant""#,
-    )
-    .unwrap();
-    let (session_id_file, session_history_path_file) = session_file_paths();
-    guest_agent::paths::write_private(&session_id_file, "partial-session").unwrap();
-    guest_agent::paths::write_private(
-        &session_history_path_file,
-        history_path.to_string_lossy().as_ref(),
-    )
-    .unwrap();
-
-    let prepare_mock = server.mock(|when, then| {
-        when.method(POST)
-            .path("/api/webhooks/agent/checkpoints/prepare-history");
-        then.status(200);
-    });
-    let checkpoint_mock = server.mock(|when, then| {
-        when.method(POST).path("/api/webhooks/agent/checkpoints");
-        then.status(200);
-    });
-
-    let result = guest_agent::checkpoint::create_recovery_checkpoint_for_runtime(&runtime).await;
-
-    let err = result.unwrap_err();
-    assert!(
-        err.to_string()
-            .contains("Session history line 2 is not valid JSON"),
-        "expected recovery checkpoint to fail on partial JSONL history, got: {err}"
-    );
-    assert!(
-        !std::path::Path::new(runtime.paths.checkpoint_error_file()).exists(),
-        "recovery checkpoint must not write the success-path checkpoint error file"
-    );
-    prepare_mock.assert_calls_async(0).await;
-    checkpoint_mock.assert_calls_async(0).await;
-}
-
-#[tokio::test]
-async fn recovery_checkpoint_rejects_non_utf8_session_history() {
-    let api = SharedApiMock::new().await;
-    let server = api.server();
-
-    let runtime = runtime_from_process_env().unwrap();
-    let _files_guard = SessionCheckpointFilesGuard::new();
-    let _history_dir = write_literal_session_history(
-        "recovery-non-utf8-session",
-        b"{\"type\":\"system\"}\nnon-utf8:\xC3(\n",
-    )
-    .unwrap();
-
-    let prepare_mock = server.mock(|when, then| {
-        when.method(POST)
-            .path("/api/webhooks/agent/checkpoints/prepare-history");
-        then.status(200);
-    });
-    let checkpoint_mock = server.mock(|when, then| {
-        when.method(POST).path("/api/webhooks/agent/checkpoints");
-        then.status(200);
-    });
-
-    let result = guest_agent::checkpoint::create_recovery_checkpoint_for_runtime(&runtime).await;
-
-    let err = result.unwrap_err();
-    assert!(
-        err.to_string()
-            .contains("Session history is not valid UTF-8"),
-        "expected recovery checkpoint to fail on invalid UTF-8 history, got: {err}"
-    );
-    assert!(
-        !std::path::Path::new(runtime.paths.checkpoint_error_file()).exists(),
-        "recovery checkpoint must not write the success-path checkpoint error file"
-    );
-    prepare_mock.assert_calls_async(0).await;
-    checkpoint_mock.assert_calls_async(0).await;
-}
-
-#[tokio::test]
-async fn recovery_checkpoint_skips_when_session_id_is_missing() {
-    let api = SharedApiMock::new().await;
-    let server = api.server();
-
-    let runtime = runtime_from_process_env().unwrap();
-    let _files_guard = SessionCheckpointFilesGuard::new();
-
-    let prepare_mock = server.mock(|when, then| {
-        when.method(POST)
-            .path("/api/webhooks/agent/checkpoints/prepare-history");
-        then.status(200);
-    });
-    let checkpoint_mock = server.mock(|when, then| {
-        when.method(POST).path("/api/webhooks/agent/checkpoints");
-        then.status(200);
-    });
-
-    let result = guest_agent::checkpoint::create_recovery_checkpoint_for_runtime(&runtime).await;
-
-    let err = result.unwrap_err();
-    assert!(
-        err.to_string().contains("No session ID found"),
-        "expected recovery checkpoint to fail on missing session ID, got: {err}"
-    );
-    assert!(
-        !std::path::Path::new(runtime.paths.checkpoint_error_file()).exists(),
-        "recovery checkpoint must not write the success-path checkpoint error file"
-    );
-    prepare_mock.assert_calls_async(0).await;
-    checkpoint_mock.assert_calls_async(0).await;
-}
-
-#[tokio::test]
-async fn recovery_checkpoint_skips_when_derived_history_is_missing() {
-    let api = SharedApiMock::new().await;
-    let server = api.server();
-
-    let runtime = runtime_from_process_env().unwrap();
-    let _files_guard = SessionCheckpointFilesGuard::new();
-    let (session_id_file, _) = session_file_paths();
-    guest_agent::paths::write_private(&session_id_file, "missing-history").unwrap();
-
-    let prepare_mock = server.mock(|when, then| {
-        when.method(POST)
-            .path("/api/webhooks/agent/checkpoints/prepare-history");
-        then.status(200);
-    });
-    let checkpoint_mock = server.mock(|when, then| {
-        when.method(POST).path("/api/webhooks/agent/checkpoints");
-        then.status(200);
-    });
-
-    let result = guest_agent::checkpoint::create_recovery_checkpoint_for_runtime(&runtime).await;
-
-    let err = result.unwrap_err();
-    assert!(
-        err.to_string().contains("Failed to read session history"),
-        "expected recovery checkpoint to fail on missing derived history, got: {err}"
-    );
-    assert!(
-        !std::path::Path::new(runtime.paths.checkpoint_error_file()).exists(),
-        "recovery checkpoint must not write the success-path checkpoint error file"
-    );
-    prepare_mock.assert_calls_async(0).await;
-    checkpoint_mock.assert_calls_async(0).await;
-}
-
-#[tokio::test]
-async fn recovery_checkpoint_rejects_invalid_session_id_without_marker() {
-    let api = SharedApiMock::new().await;
-    let server = api.server();
-
-    let runtime = runtime_from_process_env().unwrap();
-    let _files_guard = SessionCheckpointFilesGuard::new();
-    let (session_id_file, _) = session_file_paths();
-    guest_agent::paths::write_private(&session_id_file, "../unsafe-session").unwrap();
-
-    let prepare_mock = server.mock(|when, then| {
-        when.method(POST)
-            .path("/api/webhooks/agent/checkpoints/prepare-history");
-        then.status(200);
-    });
-    let checkpoint_mock = server.mock(|when, then| {
-        when.method(POST).path("/api/webhooks/agent/checkpoints");
-        then.status(200);
-    });
-
-    let result = guest_agent::checkpoint::create_recovery_checkpoint_for_runtime(&runtime).await;
-
-    let err = result.unwrap_err();
-    assert!(
-        err.to_string()
-            .contains("Failed to derive session history marker from session ID"),
-        "expected recovery checkpoint to fail on invalid session ID, got: {err}"
-    );
-    assert!(
-        !std::path::Path::new(runtime.paths.checkpoint_error_file()).exists(),
-        "recovery checkpoint must not write the success-path checkpoint error file"
-    );
-    prepare_mock.assert_calls_async(0).await;
-    checkpoint_mock.assert_calls_async(0).await;
 }
